refactor: 방 설정 적용 mutation과 핸들러를 useRoomSetting과 함께 관리 #171
