+++ frontend/src/components/common/RoomSettingModal/hooks/useRoomSetting.ts
@@ -1,10 +1,19 @@
+import { useParams } from 'react-router-dom';
+
+import useApplyRoomSetting from './useApplyRoomSetting';
 import useCategoryDropdown from './useCategoryDropdown';
 import useTimerPerRound from './useTimerPerRound';
 import useTotalRound from './useTotalRound';
 
 import { useGetRoomInfo } from '@/pages/ReadyPage/useGetRoomInfo';
 
-const useRoomSetting = () => {
+interface UseRoomSettingProps {
+  onClose: () => void;
+}
+
+const useRoomSetting = ({ onClose }: UseRoomSettingProps) => {
+  const { roomId } = useParams();
+  const { mutate: applyRoomSetting } = useApplyRoomSetting(Number(roomId));
   const { roomSetting: selectedRoomSetting } = useGetRoomInfo();
 
   const { category, handleClickOption } = useCategoryDropdown(selectedRoomSetting?.category);
@@ -13,7 +22,14 @@ const useRoomSetting = () => {
 
   const roomSetting = { category, totalRound, timerPerRound };
 
-  return { roomSetting, handleClickOption, handleClickRound, handleClickTimer };
+  const handleClickApply = () => {
+    if (!category || !totalRound || !timerPerRound) return;
+
+    applyRoomSetting({ category, totalRound, timeLimit: timerPerRound });
+    onClose();
+  };
+
+  return { roomSetting, handleClickOption, handleClickRound, handleClickTimer, handleClickApply };
 };
 
 export default useRoomSetting;
